Add files via upload
This is a simple calculator program that allows users to perform basic calculations. When the program runs, it presents the user with a menu of options:

Select operation:

Add
Subtract
Multiply
Divide
The user is prompted to enter the first number, followed by the second number. The program then performs the selected calculation based on the chosen operation. Finally, the program displays the result of the calculation.

This calculator provides a straightforward way to perform basic arithmetic operations and is a good starting point for learning about programming concepts in Python.