Add files via upload
In this code, the roll_dice function takes the number of dice and the number of sides on each die as parameters. It generates random rolls for each die and stores them in a list. The function then returns the list of rolls.

The main function prompts the user for the number of dice and the number of sides. It calls the roll_dice function to obtain the rolls and calculates the total sum of the rolls. Finally, it displays the results and the total.

You can run this code and try different combinations of dice and sides to simulate rolling the dice.