findvid2.py
In this updated code, I've added the tqdm library, which provides a simple way to create progress bars. The tqdm function is applied to the files iterable in the innermost loop, and the progress bar is updated with each iteration.

Make sure you have the tqdm library installed before running this code. You can install it using pip install tqdm.

Now, when you run the code, you will see a progress bar indicating the search progress, displaying the number of files searched and the current file being processed.