Update twitdown3.py
In this updated code, the requests.get() calls are wrapped in a try-except block. The response.raise_for_status() method is used to check for any HTTP errors. If an error occurs during the download process, it will be caught and an error message will be printed.
Remember to replace the video_url variable with the actual URL of the Twitter video you want to download.