Update donut.py
In this modified code, the donut characters are printed with the added color code \033[1;32m before each character and \033[0m after each character. This sets the color to green for the donut characters. You can modify the color code to any other ANSI color code to achieve different color effects.
Now, when you run the code, the donut characters will be displayed in green color.