Update donut.py
In  this updated code, the animation loop checks for user input using select.select() and sys.stdin.readline(). The following keys can be used to control the animation:
'q': Quit the program
'f': Increase the rotation speed in the forward direction
's': Decrease the rotation speed in the forward direction
'r': Increase the rotation speed in the reverse direction
'l': Decrease the rotation speed in the reverse direction
You can run this code and press the corresponding keys to control the animation. Press 'q' to quit the program.